Pausa pra almoço ninguem é de ferro, algumas correções no clientes e criação das funções na produtoDAO

+++ Lojinha/src/main/java/com/grupo02/lojinha/UTIL/validaTela.java
@@ -31,7 +31,7 @@ public class validaTela {
             if(isEmpty(txt))
                 throw new IllegalArgumentException();
             
-            int textValue = Integer.parseInt(txt.getText());
+            long textValue = Long.parseLong(txt.getText());
             txt.setBackground(Color.WHITE);
             return true;
         }catch(NumberFormatException e){
